# More CI optimizations