--- a/integration/invalidation/multi-level-editing/resources/build.mill
+++ b/integration/invalidation/multi-level-editing/resources/build.mill
@@ -3,9 +3,7 @@ package build
 import mill._, scalalib._
 import scalatags.Text.all._
 
-object foo extends ScalaModule {
-  def scalaVersion = "3.3.3"
-
+object foo extends JavaModule {
   def forkEnv = Map(
     "snippet" -> frag(h1("hello"), p("world"), p(constant.Constant.scalatagsVersion)).render
   )