--- a/integration/invalidation/multi-level-editing/resources/foo/src/Example.java
+++ b/integration/invalidation/multi-level-editing/resources/foo/src/Example.java
@@ -1,0 +1,8 @@
+package foo;
+
+public class Example {
+  public static void main(String[] args) {
+    String snippet = System.getenv("snippet");
+    System.out.println(snippet + "!");
+  }
+}