--- a/integration/invalidation/multi-level-editing/resources/foo/src/Example.scala
+++ b/integration/invalidation/multi-level-editing/resources/foo/src/Example.scala
@@ -1,7 +1,0 @@
-package foo
-
-object Example {
-  def main(args: Array[String]): Unit = {
-    println(sys.env("snippet") + "!")
-  }
-}