This PR removes the meta-meta-build testing from `integration/invalidation/multi-level-editing`. These tests are very slow with a 24-entry test matrix exercising 4 different kinds of edits, across 3 levels of meta-build, across both {daemon,nondaemon} modes. Meta-meta-builds are a sufficiently obscure use case we don't need to test them quite so rigorously, limiting it to 2 levels of meta-builds and trimming the matrix slightly to 16 entries

Cuts the time taken for these tests from (41s,59s,66s,87s) to (29s, 41s, 42s, 48s), which should shave about a minute off both copies (daemon, nodaemon) of the job running with 2x parallelism